@@ -3,37 +3,37 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@types/history": "^4.7.5",
-    "@types/react": "16.9.23",
-    "@types/react-beautiful-dnd": "^12.1.1",
-    "@types/react-dom": "16.9.5",
+    "@types/history": "^4.7.8",
+    "@types/react-beautiful-dnd": "^13.0.0",
+    "@types/react-dom": "^16.9.9",
     "@types/react-onclickoutside": "^6.7.3",
-    "@types/react-redux": "^7.1.7",
-    "@types/react-router": "^5.1.4",
-    "@types/react-router-dom": "^5.1.3",
+    "@types/react-redux": "^7.1.11",
+    "@types/react-router": "^5.1.8",
+    "@types/react-router-dom": "^5.1.6",
     "@types/react-timeago": "^4.1.1",
-    "@types/uuid": "^7.0.0",
-    "@types/yup": "^0.26.32",
-    "connected-react-router": "^6.7.0",
-    "formik": "^2.1.4",
+    "@types/uuid": "^8.3.0",
+    "@types/yup": "^0.29.9",
+    "connected-react-router": "^6.8.0",
+    "formik": "^2.2.5",
     "history": "^4.10.1",
-    "query-string": "^6.11.1",
-    "react": "^16.13.0",
+    "query-string": "^6.13.7",
+    "react": "^17.0.1",
     "react-beautiful-dnd": "^13.0.0",
-    "react-dom": "^16.13.0",
-    "react-markdown": "^4.3.1",
+    "react-dom": "^17.0.1",
+    "react-markdown": "^5.0.2",
     "react-onclickoutside": "^6.9.0",
-    "react-redux": "^7.2.0",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.0",
-    "react-timeago": "^4.4.0",
+    "react-redux": "^7.2.2",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "3.4.3",
+    "react-timeago": "^5.2.0",
     "redux": "^4.0.5",
     "reselect": "^4.0.0",
+    "stripe": "^8.121.0",
     "typesafe-actions": "^5.1.0",
-    "typescript": "3.8.3",
-    "uuid": "^7.0.2",
-    "yup": "^0.28.1"
+    "typescript": "4.0.5",
+    "uuid": "^8.3.1",
+    "yup": "^0.29.3"
   },
   "scripts": {
     "start": "react-scripts start",
