@@ -2,10 +2,13 @@ package main
 
 import (
 	"context"
+	"log"
 	"net/http"
 
 	"github.com/go-chi/jwtauth"
+	"github.com/go-chi/render"
 	"github.com/jmoiron/sqlx"
+	"github.com/pkg/errors"
 )
 
 // Env ...
@@ -78,7 +81,7 @@ func User() func(next http.Handler) http.Handler {
 
 			s := GetEnv(r).Service
 			_, claims, _ := jwtauth.FromContext(r.Context())
-			accountID, aok := claims.Get("id")
+			accountID, aok := claims["id"]
 
 			var acc *Account
 			if aok {
@@ -189,6 +192,101 @@ func RequireEditor() func(next http.Handler) http.Handler {
 				http.Error(w, http.StatusText(401), 401)
 				return
 			}
+			next.ServeHTTP(w, r)
+		}
+		return http.HandlerFunc(fn)
+	}
+}
+
+// RequireSubscription  ...
+func RequireSubscription() func(next http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		fn := func(w http.ResponseWriter, r *http.Request) {
+
+			s := GetEnv(r).Service.GetSubscriptionObject()
+
+			if !subscriptionIsActive(s) {
+				http.Error(w, http.StatusText(401), 401)
+				return
+			}
+			next.ServeHTTP(w, r)
+		}
+		return http.HandlerFunc(fn)
+	}
+}
+
+func subscriptionIsActive(s *Subscription) bool {
+	switch s.Status {
+
+	case "active":
+		return true
+	case "incomplete", "incomplete_expired", "past_due", "canceled":
+		return false
+	case "trialing":
+		if subHasExpired(s) {
+			return false
+		}
+		return true
+	}
+	return false
+}
+
+// RequireTrialOrPro  ...
+func RequireTrialOrPro() func(next http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		fn := func(w http.ResponseWriter, r *http.Request) {
+
+			s := GetEnv(r).Service.GetSubscriptionObject()
+
+			log.Println("level " + s.Level)
+
+			switch s.Level {
+
+			case "PRO", "TRIAL":
+				break
+			default:
+				return
+			}
+
+			next.ServeHTTP(w, r)
+		}
+		return http.HandlerFunc(fn)
+	}
+}
+
+func requireChangeableSubscription() func(next http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		fn := func(w http.ResponseWriter, r *http.Request) {
+
+			s := GetEnv(r).Service.GetSubscriptionObject()
+
+			switch s.Status {
+			case "active", "past_due":
+				break
+			default:
+				http.Error(w, http.StatusText(401), 401)
+				return
+			}
+
+			next.ServeHTTP(w, r)
+		}
+		return http.HandlerFunc(fn)
+	}
+}
+
+func requireDeleteableWorkspace() func(next http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		fn := func(w http.ResponseWriter, r *http.Request) {
+
+			s := GetEnv(r).Service.GetSubscriptionObject()
+
+			switch s.Status {
+			case "active", "past_due":
+				_ = render.Render(w, r, ErrInvalidRequest(errors.New("cannot delete workspace with an active subscription - cancel subscription first")))
+				return
+			default:
+			}
+
 			next.ServeHTTP(w, r)
 		}
 		return http.HandlerFunc(fn)
