@@ -3,6 +3,9 @@ import { loadMilestones } from '../store/milestones/actions';
 import { loadWorkflows } from '../store/workflows/actions';
 import { loadSubWorkflows } from '../store/subworkflows/actions';
 import { loadFeatures } from '../store/features/actions';
+import { loadPersonas } from '../store/personas/actions';
+import { loadWorkflowPersonas } from '../store/workflowpersonas/actions';
+import { loadFeatureComments } from '../store/featurecomments/actions';
 import { loadProjects } from '../store/projects/actions';
 import { RouteComponentProps } from 'react-router'
 import { Route, Switch, Link } from 'react-router-dom'
@@ -16,6 +19,7 @@ import { IWorkflow } from '../store/workflows/types';
 import { ISubWorkflow } from '../store/subworkflows/types';
 import { IFeature } from '../store/features/types';
 import { features } from '../store/features/selectors';
+import { featureComments, filterFeatureCommentsOnProject } from '../store/featurecomments/selectors';
 import Board from '../components/Board';
 import { filterWorkflowsOnProject, workflows } from '../store/workflows/selectors';
 import { filterMilestonesOnProject, milestones } from '../store/milestones/selectors';
@@ -24,6 +28,11 @@ import { subWorkflows } from '../store/subworkflows/selectors';
 import { Button } from '../components/elements';
 import ExternalEntityDetailsPage from './ExternalEntityDetailsPage';
 import queryString from 'query-string'
+import { IFeatureComment } from '../store/featurecomments/types';
+import { filterPersonasOnProject, personas } from '../store/personas/selectors';
+import { filterWorkflowPersonasOnProject, workflowPersonas } from '../store/workflowpersonas/selectors';
+import { IPersona } from '../store/personas/types';
+import { IWorkflowPersona } from '../store/workflowpersonas/types';
 
 
 const mapDispatchToProps = {
@@ -31,25 +40,34 @@ const mapDispatchToProps = {
     loadMilestones,
     loadWorkflows,
     loadSubWorkflows,
-    loadFeatures
+    loadFeatures,
+    loadFeatureComments,
+    loadPersonas,
+    loadWorkflowPersonas
 };
 
 const mapStateToProps = (state: AppState) => ({
     application: state.application.application,
     features: features(state),
+    featureComments: featureComments(state),
     projects: projects(state),
     milestones: milestones(state),
     workflows: workflows(state),
-    subWorkflows: subWorkflows(state)
+    subWorkflows: subWorkflows(state),
+    personas: personas(state),
+    workflowPersonas: workflowPersonas(state)
 });
 
 interface PropsFromState {
     application: IApplication
     features: IFeature[]
+    featureComments: IFeatureComment[]
     projects: IProject[]
     milestones: IMilestone[]
     workflows: IWorkflow[]
     subWorkflows: ISubWorkflow[]
+    personas: IPersona[]
+    workflowPersonas: IWorkflowPersona[]
 }
 interface RouterProps extends RouteComponentProps<{
     key: string
@@ -60,6 +78,9 @@ interface PropsFromDispatch {
     loadWorkflows: typeof loadWorkflows
     loadSubWorkflows: typeof loadSubWorkflows
     loadFeatures: typeof loadFeatures
+    loadFeatureComments: typeof loadFeatureComments
+    loadPersonas: typeof loadPersonas
+    loadWorkflowPersonas: typeof loadWorkflowPersonas
 }
 interface SelfProps { }
 type Props = RouterProps & PropsFromState & PropsFromDispatch & SelfProps
@@ -69,6 +90,7 @@ interface State {
     projectId?: string
     showClosedMilestones: boolean
     demo: boolean
+    showPersonas: boolean
 }
 
 class ExternalLinkPage extends Component<Props, State> {
@@ -78,7 +100,8 @@ class ExternalLinkPage extends Component<Props, State> {
             loading: true,
             projectId: undefined,
             showClosedMilestones: false,
-            demo: false
+            demo: false,
+            showPersonas: false
         }
     }
 
@@ -91,16 +114,20 @@ class ExternalLinkPage extends Component<Props, State> {
         API_GET_EXTERNAL_LINK(this.props.match.params.key)
             .then(response => {
                 if (!response.ok) {
-                    this.setState({loading: false})
+                    this.setState({ loading: false })
                 } else {
                     response.json().then((data: API_GET_PROJECT_RESP) => {
                         this.props.loadProjects([data.project]);
                         this.props.loadFeatures(data.features);
+                        this.props.loadFeatureComments(data.featureComments);
                         this.props.loadMilestones(data.milestones);
                         this.props.loadWorkflows(data.workflows);
                         this.props.loadSubWorkflows(data.subWorkflows);
-                        this.setState({projectId: data.project.id});
-                        this.setState({loading: false})
+                        this.props.loadPersonas(data.personas);
+                        this.props.loadWorkflowPersonas(data.workflowPersonas);
+
+                        this.setState({ projectId: data.project.id });
+                        this.setState({ loading: false })
                     })
                 }
 
@@ -133,11 +160,14 @@ class ExternalLinkPage extends Component<Props, State> {
                             </div>
                             <div className="flex items-center">
                                 <div className=" flex items-center  text-sm">
+                                    <div >
+                                        <Button title="Personas" icon="person_outline" noborder handleOnClick={() => this.setState({ showPersonas: true })} />
+                                    </div>
                                     <div>
                                         {this.state.showClosedMilestones ?
-                                            <Button iconColor="text-green-500-500" icon="toggle_on" title="Show closed" handleOnClick={() => this.setState({ showClosedMilestones: false })} />
+                                            <Button noborder iconColor="text-green-500" icon="toggle_on" title="Show closed" handleOnClick={() => this.setState({ showClosedMilestones: false })} />
                                             :
-                                            <Button icon="toggle_off " title="Show closed" handleOnClick={() => this.setState({ showClosedMilestones: true })} />
+                                            <Button noborder icon="toggle_off " title="Show closed" handleOnClick={() => this.setState({ showClosedMilestones: true })} />
                                         }
                                     </div>
                                 </div>
@@ -157,6 +187,13 @@ class ExternalLinkPage extends Component<Props, State> {
                                 projectId={project.id}
                                 workspaceId={project.workspaceId}
                                 demo={this.state.demo}
+                                comments={filterFeatureCommentsOnProject(this.props.featureComments, project.id)}
+                                personas={filterPersonasOnProject(this.props.personas, project.id)}
+                                workflowPersonas={filterWorkflowPersonasOnProject(this.props.workflowPersonas, project.id)}
+
+                                showPersonas={this.state.showPersonas}
+                                closePersonas={() => this.setState({ showPersonas: false })}
+                                openPersonas={() => this.setState({ showPersonas: true })}
                             />
                         </div>
                     </div>
