@@ -96,15 +96,15 @@ class SignUp extends Component<Props> {
 
 
                                 actions.setSubmitting(false)
-                            }}                        
+                            }}
                         >
-                        {(formikBag: FormikProps<API_SIGN_UP_REQ>) => (
+                            {(formikBag: FormikProps<API_SIGN_UP_REQ>) => (
                                 <Form>
                                     {formikBag.status && formikBag.status.msg && <div>{formikBag.status.msg}</div>}
                                     <Field
-                                        name="workspaceName"                                        
+                                        name="workspaceName"
                                     >
-                                    {({ field, form }: FieldProps<API_SIGN_UP_REQ>) => (
+                                        {({ field, form }: FieldProps<API_SIGN_UP_REQ>) => (
                                             <div className="flex flex-col    items-baseline sm:flex-row">
                                                 <div className=" flex flex-col w-full">
                                                     <div><input type="text" value={form.values.workspace} onChange={form.handleChange} placeholder="workspace name" id="workspaceName" className="rounded p-2 border w-full text-lg	" /></div>
@@ -112,21 +112,21 @@ class SignUp extends Component<Props> {
                                                 </div>
                                             </div>
                                         )}
-                                        </Field>
+                                    </Field>
                                     <Field
-                                        name="name"                                        
+                                        name="name"
                                     >
-                                    {({ field, form }: FieldProps<API_SIGN_UP_REQ>) => (
+                                        {({ field, form }: FieldProps<API_SIGN_UP_REQ>) => (
                                             <div className="flex  flex-row items-baseline">
                                                 <div className=" flex flex-col w-full">
-                                                    <div><input type="text" value={form.values.name} onChange={form.handleChange}  placeholder="Name, e.g. John Smith" id="name" className=" w-full rounded p-2 border  text-lg	" /></div>
+                                                    <div><input type="text" value={form.values.name} onChange={form.handleChange} placeholder="Name, e.g. John Smith" id="name" className=" w-full rounded p-2 border  text-lg	" /></div>
                                                     <div className="m-1 text-red-500 text-xs font-bold">{form.touched.name && form.errors.name && form.errors.name}</div>
                                                 </div>
                                             </div>
                                         )}
-                                        </Field>
+                                    </Field>
                                     <Field
-                                        name="email"                                        
+                                        name="email"
                                     >
                                         {({ field, form }: FieldProps<API_SIGN_UP_REQ>) => (
                                             <div className="flex  flex-row items-baseline">
@@ -137,11 +137,11 @@ class SignUp extends Component<Props> {
                                                 </div>
                                             </div>
                                         )}
-                                                                        </Field>
+                                    </Field>
                                     <Field
                                         name="password"
                                     >
-                                    {({ field, form }: FieldProps<API_SIGN_UP_REQ>) => (
+                                        {({ field, form }: FieldProps<API_SIGN_UP_REQ>) => (
                                             <div className="flex flex-row items-baseline">
                                                 <div className="flex flex-col w-full">
                                                     <div><input type="password" value={form.values.password} onChange={form.handleChange} placeholder="Password" id="password" className="rounded p-2 border w-full text-lg	" /></div>
@@ -149,7 +149,7 @@ class SignUp extends Component<Props> {
                                                 </div>
                                             </div>
                                         )}
-                                        </Field>
+                                    </Field>
                                     <div className="flex justify-center">
                                         <Button submit primary title="Create account" />
                                     </div>
