@@ -6,7 +6,9 @@ export enum ActionTypes {
     LOAD_FEATURES = 'LOAD_FEATURES',
     UPDATE_FEATURE = 'UPDATE_FEATURE',
     DELETE_FEATURE = 'DELETE_FEATURE',
-    MOVE_FEATURE = 'MOVE_FEATURE'
+    MOVE_FEATURE = 'MOVE_FEATURE',
+    DELETE_FEATURES_BY_MILESTONE = 'DELETE_FEATURES_BY_MILESTONE',
+    DELETE_FEATURES_BY_SUBWORKFLOW = 'DELETE_FEATURES_BY_SUBWORKFLOW',
 }
 
 export interface createFeature { type: ActionTypes.CREATE_FEATURE, payload: IFeature }
@@ -21,8 +23,15 @@ export const updateFeature = (x: IFeature) => action(ActionTypes.UPDATE_FEATURE,
 export interface deleteFeature { type: ActionTypes.DELETE_FEATURE, payload: string }
 export const deleteFeature = (x: string) => action(ActionTypes.DELETE_FEATURE, x)
 
+
+export interface deleteAllFeaturesByMilestone { type: ActionTypes.DELETE_FEATURES_BY_MILESTONE, payload: string }
+export const deleteAllFeaturesByMilestone = (x: string) => action(ActionTypes.DELETE_FEATURES_BY_MILESTONE, x)
+
+export interface deleteAllFeaturesBySubWorkflow { type: ActionTypes.DELETE_FEATURES_BY_SUBWORKFLOW, payload: string }
+export const deleteAllFeaturesBySubWorkflow = (x: string) => action(ActionTypes.DELETE_FEATURES_BY_SUBWORKFLOW, x)
+
 interface moveFeaturePayload { id: string, toMilestoneId: string, toSubWorkflowId: string, index: number, ts: string, by: string }
 export interface moveFeature { type: ActionTypes.MOVE_FEATURE, payload: moveFeaturePayload }
 export const moveFeature = (x: moveFeaturePayload) => action(ActionTypes.MOVE_FEATURE, x)
 
-export type Actions = createFeature | loadFeatures | updateFeature | deleteFeature | moveFeature
+export type Actions = createFeature | loadFeatures | updateFeature | deleteFeature | moveFeature | deleteAllFeaturesByMilestone | deleteAllFeaturesBySubWorkflow
